requirements.txt
aiohttp==3.6.3
akinator.py==4.1.0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
discord==1.0.1
discord.py==1.5.1
dnspython==2.1.0
Flask==1.1.2
giphy-client==1.0.0
google==3.0.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imgurpython==1.1.7
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
more-itertools==8.7.0
multidict==4.7.6
Pillow==8.0.1
praw==7.1.0
prawcore==1.5.0
pycparser==2.20
pyfiglet==0.8.post1
pymongo==3.11.3
PyNaCl==1.4.0
python-dateutil==2.8.1
pytube==10.0.0
requests==2.25.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
Werkzeug==1.0.1
wikipedia==1.4.0
wolframalpha==4.1.1
xmltodict==0.12.0
yarl==1.5.1
youtube-dl==2020.12.9
